updates to routes

+++ routes/sheets.js
@@ -3,11 +3,11 @@ const router = require("express").Router();
 
 router.post("/createsheet", function(req, res) {
 	// console.log("inside create sheet route");
-	// console.log(req.user)
  	
 	let newSheetName = req.body.newSheetName;
 	let userId = req.user.id;
 
+	//create a row in the Sheet table with the name of the new sheet
  	db.Sheet.create({
  		name: newSheetName
  	}).then(function(result) {
@@ -15,6 +15,7 @@ router.post("/createsheet", function(req, res) {
 	 	let newSheetData = result.dataValues;
 		console.log("new sheet id is " + newSheetData.id);
 
+		//Create a row in the UserSheet table which records the id of the user, the id of the newly created sheet, and lists the user as the creator of this sheet
 		db.UserSheet.create({
 			userId: userId,
 			sheetId: newSheetData.id,
@@ -44,6 +45,7 @@ router.post("/createsheet", function(req, res) {
 router.get("/viewcreated", function(req, res) {
 	let userId = req.user.id;
 
+	//finds all the sheets that the user himself as created, along with any collaborators
 	db.UserSheet.findAll({
 		include: [{
 			model: db.Sheet,
@@ -76,6 +78,7 @@ router.get("/viewcreated", function(req, res) {
 
 router.get("/viewshared", function(req, res) {
 	let userId = req.user.id;
+	//finds all of the user's shared sheets that he didn't create but has access to, along with any collaborators
  	// console.log(req.body)
  	db.UserSheet.findAll({
 		include: [{
@@ -216,7 +219,7 @@ router.post("/createtransaction", function(req, res) {
 	// 	"departmentName": "Sales",
 	// 	"locationName": "City of choice",
 	// 	"representativeName": "Bobby Jr",
-	// 	"SheetId": 1
+	// 	"sheetId": 1
 	// }
 
 	//*When this is really working, companyName will be equivalent to the name of the user who is creating the transaction. So something like:
@@ -245,18 +248,57 @@ router.post("/createtransaction", function(req, res) {
 	});
 });
 
-router.delete("/deletetransaction/:sheetRecords/:userSheet", function(req, res) {
-	console.log(req.body)
+router.delete("/deletetransaction/:sheetId/:transactionId/:userId", function(req, res) {
 	
-	let transaction = req.body.id;
+	//Check to see if the user is signed in with the same userid in the URL
+	let sessionUserId = req.user.id;//*make sure this works later on*
+	let requestUserId = req.params.userId;
 
-	db.Transaction.destroy({
+	if (sessionUserId !== requestUserId) {
+		console.log('user ids do not match')
+		res.status(401).send("session user id and request user id do not match");
+		return;
+	}
+
+	let transactionId = req.params.transactionId;
+	let sheetId = req.params.sheetId;
+
+	//check to see if the user has access to the sheet the transaction is in
+	db.UserSheet.findOne({
 		where: {
-			id: req.params.id,
+			sheetId: sheetId,
+			userId: sessionUserId
 		}
 	}).then(function(result) {
-		res.json(dbTransaction)
-	}).catch(function (err) {
+
+		//check to see if the user himself created the transaction in question
+		db.Transaction.findOne({
+			where: {
+				companyName: req.user.name,
+				id: transactionId
+			}
+		}).then(function(result) {
+			
+			//delete the transaction
+			db.Transaction.destroy({
+				where: {
+					id: transactionId
+				}
+			}).then(function(result) {
+				console.log('transaction deleted with id' + req)
+				res.status(200).send("transaction deleted")
+			}).catch(function (err) {
+				console.log(err);
+				res.json(err);
+			});
+
+		}).catch(function(err) {
+			console.log('either the user did not create this transaction, or the transaction under this id does not exist');
+			res.json(err)
+		})
+		
+	}).catch(function(err) {
+		console.log('the user does not have access to the sheet this transaction is in')
 		console.log(err);
 		res.json(err);
 	});
@@ -379,23 +421,68 @@ router.get("/getSheetData/:sheetId", function(req, res) {
 	});
 })
 
-router.delete("/deletechart/:userName/:userEmail/:userId/:transactions/:collaborators", function(req, res) {
-	console.log(req.body)
-	
-	db.Chart.destroy({
+router.delete("/deletesheet/:sheetId/:userId/", function(req, res) {
+	//console.log(req.body)
+	let sessionUserId = req.body.id;//*make sure this works later on*
+	let requestUserId = req.params.userId;
+
+	if (sessionUserId !== requestUserId) {
+		console.log('user ids do not match')
+		res.status(401).send("session user id and request user id do not match");
+		return;
+	}
+
+	let sheetId = req.params.sheetId;
+
+	//check to see if user is the creator of this sheet and has permission to delete it
+	db.UserSheet.findOne({
 		where: {
-			userName: req.body.userName,
-			userEmail: req.body.userEmail,
-			userId: req.body.userId,
-			transactions: req.body.transactions,
-			collaborators: req.body.vendorNumber
+			sheetId: sheetId,
+			userId: req.params.userId,
+			userIsCreator: true
 		}
 	}).then(function(result) {
-		res.json(dbChart)
-	}).catch(function (err) {
-			console.log(err);
+
+		//delete all usersheet records associated with this sheet
+		db.UserSheet.destroy({
+			where: {
+				sheetId: sheetId
+			}
+		}).then(function(result) {
+			//delete the sheet table record with the sheetId in question
+			db.Sheet.destroy({
+				where: {
+					id: sheetId
+				}
+			}).then(function(result) {
+
+				db.Transaction.destroy({
+					where: {
+						sheetId: sheetId
+					}
+				}).then(function(result) {
+					console.log('all records for this sheet deleted from usersheets, transactions, and sheets tables')
+					res.status(200).send('all records for this sheet deleted from usersheets, transactions, and sheets tables')
+				}).catch(function(err) {
+					console.log(err)
+					res.json(err);
+				})
+
+			}).catch(function(err) {
+			console.log(err)
 			res.json(err);
-		});
+		})
+
+		}).catch(function(err) {
+			console.log(err)
+			res.json(err);
+		})
+
+	}).catch(function(err) {
+		console.log("user is not the creator of this sheet and does not have permission to delete it")
+		console.log(err)
+		res.json(err);
+	})
 });
 module.exports = router;
